+++ hydra-node/src/Hydra/Options.hs
@@ -881,7 +881,7 @@ hydraNodeCommand =
   info
     ( commandParser
         <**> versionInfo
-        <**> scriptInfo
+        <**> hydraScriptCatalogue
         <**> helper
     )
     ( fullDesc
@@ -895,11 +895,11 @@ hydraNodeCommand =
       (showVersion hydraNodeVersion)
       (long "version" <> help "Show version")
 
-  scriptInfo :: Parser (a -> a)
-  scriptInfo =
+  hydraScriptCatalogue :: Parser (a -> a)
+  hydraScriptCatalogue =
     infoOption
-      (decodeUtf8 $ encodePretty Contract.scriptInfo)
-      (long "script-info" <> help "Dump script info as JSON")
+      (decodeUtf8 $ encodePretty Contract.hydraScriptCatalogue)
+      (long "hydra-script-catalogue" <> help "Dump Hydra script catalogue as JSON")
 
 defaultContestationPeriod :: ContestationPeriod
 defaultContestationPeriod = 600