+++ CHANGELOG.md
@@ -26,6 +26,8 @@ changes.
 
 - Tested with `cardano-node 10.4.1` and `cardano-cli 10.8.0.0`.
 
+- Rename `--script-info` option to `--hydra-script-catalogue`.
+
 Fix rotation log id consistency after restart by changing the rotation check to trigger only
 when the number of persisted `StateChanged` events exceeds the configured `--persistence-rotate-after` threshold.
   * This also prevents immediate rotation on startup when the threshold is set to 1.